+++ Exercises.sln
@@ -5,7 +5,11 @@ VisualStudioVersion = 17.5.33530.505
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Exercises", "Exercises.csproj", "{7E96383C-DA2F-48F8-B678-6902EFB83B28}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OOP", "..\OOP\OOP.csproj", "{5162F244-DC33-4217-A82C-B3A5B392CEC8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OOP", "..\OOP\OOP.csproj", "{5162F244-DC33-4217-A82C-B3A5B392CEC8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpriteOut", "..\SpritesOut\SpriteOut.csproj", "{013C1D2A-B814-44C5-89D6-D6733A09C225}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dictionary", "..\Dictionary\Dictionary.csproj", "{5078FA3D-FB79-45B9-ACA7-7B043D65723C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +25,14 @@ Global
 		{5162F244-DC33-4217-A82C-B3A5B392CEC8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5162F244-DC33-4217-A82C-B3A5B392CEC8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5162F244-DC33-4217-A82C-B3A5B392CEC8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{013C1D2A-B814-44C5-89D6-D6733A09C225}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{013C1D2A-B814-44C5-89D6-D6733A09C225}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{013C1D2A-B814-44C5-89D6-D6733A09C225}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{013C1D2A-B814-44C5-89D6-D6733A09C225}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5078FA3D-FB79-45B9-ACA7-7B043D65723C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5078FA3D-FB79-45B9-ACA7-7B043D65723C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5078FA3D-FB79-45B9-ACA7-7B043D65723C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5078FA3D-FB79-45B9-ACA7-7B043D65723C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
